Fix lists by boxing and aligning them to the left
List were centered on the page along with everything else. Presumably due
to the points/numbers that ul/ol creates list's text wasn't aligned with everything else.
Fix it by containing inside a bordered box that is as wide as the page content
and aligning it to the left there for better UX readability.